added room assignment automated email for locked-in rooms
