+++ hotel/config.py
@@ -2,16 +2,35 @@ from hotel import *
 
 hotel_config = parse_config(__file__)
 
+c.NIGHT_NAMES = [name.lower() for name in c.NIGHT_VARS]
 c.NIGHT_DISPLAY_ORDER = [getattr(c, night.upper()) for night in hotel_config['night_display_order']]
-c.NIGHT_NAMES = ['monday', 'tuesday', 'wednesday', 'thursday', 'friday', 'saturday', 'sunday']
+
+c.NIGHT_DATES = {c.ESCHATON.strftime('%A'): c.ESCHATON.date()}
 
 c.CORE_NIGHTS = []
 _day = c.EPOCH
 while _day.date() != c.ESCHATON.date():
+    c.NIGHT_DATES[_day.strftime('%A')] = _day.date()
     c.CORE_NIGHTS.append(getattr(c, _day.strftime('%A').upper()))
     _day += timedelta(days=1)
+
+for _before in range(1, 4):
+    _day = c.EPOCH.date() - timedelta(days=_before)
+    c.NIGHT_DATES[_day.strftime('%A')] = _day
 
 c.SETUP_NIGHTS = c.NIGHT_DISPLAY_ORDER[:c.NIGHT_DISPLAY_ORDER.index(c.CORE_NIGHTS[0])]
 c.TEARDOWN_NIGHTS = c.NIGHT_DISPLAY_ORDER[1 + c.NIGHT_DISPLAY_ORDER.index(c.CORE_NIGHTS[-1]):]
 
-c.ROOMS_LOCKED_IN = hotel_config['rooms_locked_in']
+for _attr in ['CORE_NIGHT', 'SETUP_NIGHT', 'TEARDOWN_NIGHT']:
+    setattr(c, _attr + '_NAMES', [c.NIGHTS[night] for night in getattr(c, _attr + 'S')])
+
+c.CHECK_IN_TIME = hotel_config['check_in_time']
+c.CHECK_OUT_TIME = hotel_config['check_out_time']
+
+
+@Config.mixin
+class ExtraConfig:
+    @property
+    def ONE_WEEK_OR_TAKEDOWN(self):
+        week_from_now = datetime.combine(date.today() + timedelta(days=7), time(23, 59)).replace(tzinfo=c.EVENT_TIMEZONE)
+        return min(week_from_now, c.UBER_TAKEDOWN)
