+++ hotel/configspec.ini
@@ -1,14 +1,14 @@
-
-# Hotel rooms may no longer be edited after this is True EXCEPT by admins with
-# STAFF_ROOMS access, to limit the post-deadline changes.  There's a separate
-# ROOM_DEADLINE option, which controls how long volunteers can request/decline
-# room space.  The actual deadline after which it's too late for even admins
-# to edit room information on the backend comes later, and is set manually
-# since it typically varies and is not known exactly in advance, so we use
-# this boolean instead of a date.
-rooms_locked_in = boolean(default=False)
+# MAGFest provides staff rooms for returning volunteers.  In addition to the
+# config options defined here, you must add a "room_deadline" setting to the
+# [dates] section of the main repo's config when including this plugin.
 
 # In some of our pages and emails relating to hotel room nights, it makes sense
 # to list the nights in order based on the start of the event rather than the
 # first day of the week.
 night_display_order = string_list(default=list("monday", "tuesday", "wednesday", "thursday", "friday", "saturday", "sunday"))
+
+# When people are able to pick up their room keys from the hotel front desk.
+check_in_time = string(default='4pm')
+
+# What time in the morning people need to check out of their hotel rooms.
+check_out_time = string(default='10am')
