+++ hotel/templates/hotel_assignments/index.html
